// dict/api/contacts/types.ts
/* eslint-disable */
// tslint:disable
/**
 * contact_information
 * 
 *
 * OpenAPI spec version: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface ContactsApi {
	createContact: (request: Api.BODYDATA | undefined) => Promise<CreateContactResponse>
	deleteContact: (contactId: string) => Promise<DeleteContactResponse>
	geTContact: () => Promise<GeTContactResponse>
	updateContact: (contactId: string, request: Api.BODYDATA | undefined) => Promise<UpdateContactResponse>
}

export type CreateContactResponse = CreateContact201Response | CreateContact400Response

export interface CreateContact201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface CreateContact400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type DeleteContactResponse = DeleteContact201Response | DeleteContact400Response

export interface DeleteContact201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface DeleteContact400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GeTContactResponse = GeTContact201Response | GeTContact400Response

export interface GeTContact201Response {
	status: 201
	body: Api.BODYDATA[]
	headers?: never
}

export interface GeTContact400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type UpdateContactResponse = UpdateContact201Response | UpdateContact400Response

export interface UpdateContact201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface UpdateContact400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}